비로그인 로그인 수정

+++ controllers/board-controller.js
@@ -1,10 +1,11 @@
 const { Board, Like, sequelize, Sequelize } = require('../models');
 const { Op } = require('sequelize');
 const {
-  situationBoardLogin,
-  situationBoard,
-  detailComments,
-  detailBoard,
+  situationBoardLogin, //로그인시
+  situationBoard, //비로그인
+  detailCommentsAll, //로그인비로그인시 댓글
+  detailBoard, //비로그인 상세 게시글
+  detailBoardLogin, //로그인 상세 게시글
   PopBoardHome,
   NewBoardHome,
   NewBoardHomeLogin,
@@ -14,7 +15,6 @@ const {
 const getBoardHome = async (req, res) => {
   try {
     const user_id = req.userId;
-    // const user_id = 1; //임의로 설정
     console.log('유저로그인', user_id);
 
     if (user_id) {
@@ -41,7 +41,8 @@ const getBoardHome = async (req, res) => {
 //상황별 페이지 게시글 전체 조회
 const getSituationBoard = async (req, res) => {
   try {
-    const user_id = 1; //임의로 설정
+    const user_id = req.userId;
+    console.log('유저로그인', user_id);
 
     if (user_id) {
       const board_list = await situationBoardLogin(user_id);
@@ -63,7 +64,9 @@ const getSituationBoard = async (req, res) => {
 // 게시글 디테일페이지 조회
 const getDetailBoard = async (req, res) => {
   try {
-    const user_id = 1; //임의로 설정
+    const user_id = req.userId;
+    console.log('유저로그인', user_id);
+
     const { board_id } = req.params;
 
     if (req.cookies['f' + board_id] == undefined) {
@@ -73,10 +76,18 @@ const getDetailBoard = async (req, res) => {
       });
       await Board.increment({ view_count: +1 }, { where: { board_id } });
     }
-    const result = await detailBoard(user_id, board_id);
-    const comments = await detailComments(board_id);
-    const board = result[0];
-    res.status(200).json({ result: 'success', board, comments });
+
+    if (user_id) {
+      const result = await detailBoardLogin(user_id, board_id);
+      const comments = await detailCommentsAll(board_id);
+      const board = result[0];
+      res.status(200).json({ result: 'success', board, comments });
+    } else {
+      const result = await detailBoard(board_id);
+      const comments = await detailCommentsAll(board_id);
+      const board = result[0];
+      res.status(200).json({ result: 'success', board, comments });
+    }
   } catch (error) {
     console.error(error);
     res.status(400).json(() => {
@@ -93,28 +104,32 @@ function getUserIP(req) {
 //좋아요 /좋아요 취소
 const changeLike = async (req, res) => {
   try {
-    const user_id = 1;
+    const user_id = req.userId;
     const { board_id } = req.params;
-    console.log(board_id);
-    const isLike = await Like.findOne({
-      where: {
-        [Op.and]: { board_id: board_id, user_id: user_id },
-      },
-    });
-    //isLike 가 있으면 delete 없으면 create
-    if (isLike) {
-      await Like.destroy({
+    console.log('유저로그인', user_id);
+    if (user_id) {
+      const isLike = await Like.findOne({
         where: {
           [Op.and]: { board_id: board_id, user_id: user_id },
         },
       });
-      res.status(200).json({ result: 'success', like_state: false });
+      //isLike 가 있으면 delete 없으면 create
+      if (isLike) {
+        await Like.destroy({
+          where: {
+            [Op.and]: { board_id: board_id, user_id: user_id },
+          },
+        });
+        res.status(200).json({ result: 'success', like_state: false });
+      } else {
+        await Like.create({
+          user_id: user_id,
+          board_id: board_id,
+        });
+        res.status(200).json({ result: 'success', like_state: true });
+      }
     } else {
-      await Like.create({
-        user_id: user_id,
-        board_id: board_id,
-      });
-      res.status(200).json({ result: 'success', like_state: true });
+      res.status(400).json({ result: 'fail', msg: '로그인이 필요합니다.' });
     }
   } catch (error) {
     console.error(error);
